feat(alerts+funding): unify /alerts onto the Clerk surface; first UI for the funding-repo matcher
Wave 7 of the plan — complete the Pro promise:

ALERTS UNIFICATION (7a):
- /alerts becomes a server gate: Clerk-signed-in visitors are redirected
  to /you/alerts (the canonical surface — email + webhook delivery, tier
  quotas, quiet hours); anonymous visitors keep the cookie inbox exactly
  as before (moved verbatim to AlertsClient.tsx) plus an upsell line
  explaining what an account adds. Pre-fix, a signed-in user landing on
  /alerts silently managed a SECOND, inferior rule set that their paid
  tier never applied to.
- /alerts joins the middleware isClerkSessionRoute list so the gate has
  Clerk context; without keys (CI/local) it degrades to the anonymous
  view. Deferred (documented): one-time import of anonymous cookie rules
  into the account surface.

FUNDING-REPO CROSS-LINK (7b):
- New RepoFundingStrip on /repo/[owner]/[name]: the funding→repo matcher
  (getFundingEventsForRepo — domain > alias > company-name > fuzzy,
  memoized) was fully built and consumed NOWHERE in the UI. The strip
  shows the latest matched round (company, amount, round, match reason +
  confidence) and links into /funding. Renders null for repos with no
  matched events — the vast majority.

Gate: typecheck, lint 0 errors, guards, full node:test suite, 347
vitest, production build all green.
